Move Android ICU bootstrap into icu.net with optional net10.0-android TFM.
Wrapper.Init() now configures bundled ICU on Android automatically, removing the test-app setup hook.
