+++ scripts/run-android-tests.ps1
@@ -54,5 +54,7 @@ Write-Host "Connected devices:"
 $devices | ForEach-Object { Write-Host "  $_" }
 
 Write-Host "Running Android device tests ($Configuration)..."
-dotnet test $project -f net10.0-android -c $Configuration --logger "trx;LogFileName=test-results.trx"
+$env:IcuDotNetIncludeAndroid = 'true'
+dotnet restore $project -p:IcuDotNetIncludeAndroid=true
+dotnet test $project -f net10.0-android -c $Configuration -p:IcuDotNetIncludeAndroid=true --no-restore --logger "trx;LogFileName=test-results.trx"
 exit $LASTEXITCODE
