+++ source/Directory.Build.props
@@ -19,6 +19,8 @@ See full changelog at https://github.com/sillsdev/icu-dotnet/blob/master/CHANGEL
     <SymbolPackageFormat>snupkg</SymbolPackageFormat>
     <LangVersion>default</LangVersion>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
+    <!-- Honor -p:IcuDotNetIncludeAndroid=true and the env var set by scripts/IDE before restore/build -->
+    <IcuDotNetIncludeAndroid Condition="'$(IcuDotNetIncludeAndroid)' == '' And '$([System.Environment]::GetEnvironmentVariable(`IcuDotNetIncludeAndroid`))' == 'true'">true</IcuDotNetIncludeAndroid>
     <SignAssembly>true</SignAssembly>
     <AssemblyOriginatorKeyFile>$(MSBuildThisFileDirectory)/icu.net.snk</AssemblyOriginatorKeyFile>
   </PropertyGroup>
