+++ source/icu.net.tests/icu.net.tests.csproj
@@ -12,6 +12,7 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(TargetFramework)' == 'net10.0-android'">
     <SupportedOSPlatformVersion>21.0</SupportedOSPlatformVersion>
+    <IcuDotNetIncludeAndroid>true</IcuDotNetIncludeAndroid>
   </PropertyGroup>
   <ItemGroup Condition="'$(TargetFramework)' != 'net10.0-android'">
     <PackageReference Include="Icu4c.Win.Full.Lib" Version="62.1.4-beta" />
