+++ source/icu.net/IcuWrapper.cs
@@ -95,41 +95,6 @@ namespace Icu
 			NativeMethods.PreferredDirectory = directory;
 		}
 
-		/// <summary>
-		/// Android APKs require ICU native libraries to be named libicuuc.so (without a version suffix).
-		/// Call this before <see cref="Init"/> with the major ICU version that was bundled.
-		/// </summary>
-		[PublicAPI]
-		public static void SetAndroidBundledIcuMajorVersion(int majorVersion)
-		{
-			NativeMethods.AndroidBundledIcuMajorVersion = majorVersion;
-		}
-
-		/// <summary>
-		/// Android hosts native libraries inside APK splits where dlopen needs a zip path.
-		/// Provide a loader that returns a dlopen handle for a file name such as libicuuc.so.
-		/// </summary>
-		[PublicAPI]
-		public static void SetAndroidNativeLibraryLoader(Func<string, IntPtr> loader)
-		{
-			NativeMethods.AndroidLoadNativeLibrary = loader;
-		}
-
-		/// <summary>
-		/// Mono Android FastDev cannot resolve libc P/Invokes; provide dlsym lookup for ICU entry points.
-		/// </summary>
-		[PublicAPI]
-		public static void SetAndroidSymbolResolver(Func<string, IntPtr> resolver)
-		{
-			NativeMethods.AndroidResolveSymbol = resolver;
-		}
-
-		/// <summary>
-		/// Optional hook for Android device test hosts to configure bundled ICU before NUnit runs.
-		/// </summary>
-		[PublicAPI]
-		public static Action AndroidTestConfigure { get; set; }
-
 		#region Public wrappers around the ICU methods
 
 		/// ------------------------------------------------------------------------------------
@@ -142,6 +107,9 @@ namespace Icu
 		[PublicAPI]
 		public static ErrorCode Init()
 		{
+#if __ANDROID__
+			AndroidIcuBootstrap.EnsureConfigured();
+#endif
 			NativeMethods.u_init(out var errorCode);
 			return errorCode;
 		}
